third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/object-src-image-url-blocked-expected.txt

CONSOLE ERROR: Refused to load plugin data from 'http://127.0.0.1:8000/resources/square20.jpg' because it violates the following Content Security Policy directive: "object-src 'none'".

This test passes if there is a console message saying the plugin was blocked.
